refactor: restructure chart, event and customSVGOptions dirs
