+++ src/visualizations/config/adapters/dhis_highcharts/customSVGOptions/index.js
@@ -1,18 +1,7 @@
 import { VIS_TYPE_SINGLE_VALUE } from '../../../../../modules/visTypes.js'
-import { getSingleValueCustomSVGOptions } from './singleValue/index.js'
-import { renderSingleValueSVG } from './singleValue/renderer/renderSingleValueSVG.js'
+import getSingleValueCustomSVGOptions from './singleValue/index.js'
 
-export function renderCustomSVG(visType) {
-    switch (visType) {
-        case VIS_TYPE_SINGLE_VALUE:
-            renderSingleValueSVG.call(this)
-            break
-        default:
-            break
-    }
-}
-
-export function getCustomSVGOptions({
+export default function getCustomSVGOptions({
     extraConfig,
     layout,
     extraOptions,
