+++ src/visualizations/config/adapters/dhis_highcharts/events/loadCustomSVG/singleValue/styles.js
