+++ src/visualizations/config/adapters/dhis_highcharts/index.js
@@ -14,8 +14,8 @@ import {
 } from '../../../../modules/visTypes.js'
 import { defaultMultiAxisTheme1 } from '../../../util/colors/themes.js'
 import addTrendLines, { isRegressionIneligible } from './addTrendLines.js'
-import getChart from './chart.js'
-import { getCustomSVGOptions } from './customSVGOptions/index.js'
+import getChart from './chart/index.js'
+import getCustomSVGOptions from './customSVGOptions/index.js'
 import getExporting from './exporting.js'
 import getScatterData from './getScatterData.js'
 import getSortedConfig from './getSortedConfig.js'
