+++ src/visualizations/config/adapters/dhis_highcharts/title/singleValue.js
@@ -1,5 +1,5 @@
 import getFilterText from '../../../../util/getFilterText.js'
-export { getSingleValueTitleColor } from '../customSVGOptions/singleValue/index.js'
+export { getSingleValueTitleColor } from '../customSVGOptions/singleValue/getSingleValueTitleColor.js'
 
 export function getSingleValueTitleText(layout, metaData) {
     if (layout.hideTitle) {
